+++ Changelog.md
@@ -1,5 +1,6 @@
 ## 0.0.25
 + Removed `AccessMode.Protected` access mode. Use `AccessMode.Public` or `AccessMode.Private` instead.
++ Removed `PredictionResource.id` field as it is no longer used.
 
 ## 0.0.24
 + Added `PredictionResource.type` field for inspecting the type of a prediction resource.